WeatherView xib 연결
